Tested RewardValue class
Converted the miles value to an integer o be consistent with other parts of the already existing program.

+++ src/main/java/RewardValue.java
@@ -7,7 +7,7 @@ public class RewardValue {
         this.miles =  (this.cash / milesToCashRate);
     }
 
-    RewardValue(float miles)
+    RewardValue(int miles)
     {
         this.miles = miles;
         this.cash = (milesToCashRate * this.miles);
